Simplify PPPoE credentials for KPN
Installed the 1.11.x firmware a few days ago on my UDM Pro and today installed your IGMP proxy script to finally be able to remove the KPN modem from my setup. It worked flawlessly! The instructions can be simplified a bit however, because as far as I understand it KPN allows any username/password combination for PPPoE. But as an example they state to use internet/internet. Tried that, and it works. Easier than having to look up / type the MAC address of your modem.

Thank you for your efforts!