Create Debian package for helper tool
This change adds the configuration for building a Debian package for
udm-iptv. This package installs the udm-iptv service for the target
environment, which sets up the VLAN interface, DHCP, and igmpproxy
necessary for IPTV. Configuration is located at /etc/udm-iptv.conf.

By providing a Debian package, we enable installation of this package
for the new UniFi firmware that is based on Debian and is used by the
UDM Pro SE and the UDR.