Add support for UniFi Dream Router 7 (#382)
This commit extends the installer with support for
UniFi Dream Router 7 (UDR7). It supports 3 ports to 
be configured for either WAN 1 or WAN 2.  The
ports are eth2, eth3 and eth4:
- eth2 is switch port 3 (RJ45)
- eth3 is switch port 4 (RJ45)
- eth4 is port 5 (SFP+)

In addition, this commit also introduce a new profile
entry for Freedom Internet, re-using the existing
profile config for KPN.